optimization on timeframes/index
Because we're enabling/disabling options based on each individual task,
as-is every time anything changes on `timeframes/index` we need to
re-render every task. This makes it so that all tasks show the same
thing until you actually click on the the dropdown, at which point it
calculates which options are valid, as well as the counts for each.