add unique constraint on stats
This was causing some breakage in production where if the first couple
of tasks for the day are completed in rapid succession, it might slip
past the activerecord validation and leave two records in the database.
Subsequent calls cause errors, since the records that already exist are
invalid. Preventing this at the database level seems preferable. I'm
expecting to have to clean up a couple of records before the migration
will run cleanly, though...