Note to self
It was `application.js`. I think the problem was that we're using es6
syntax in the top level file (formerly `application.js`). Previously
this was being precompiled via browserify and sprockets, but removing
browserify caused sprockets to barf. Sprockets probably somewhere
automatically adds `application.js` to its list of things to process, so
we just take that file away so it has nothing to do.