Redux: fixes for issues introduced around timeframes
Switching to 1 argument in action creator broke task updates on task
list pages, so switched back to 2 after some more thought. Seems to make
sense to keep the id separate.

Also had to fetch tasks when we click `Refresh` on the timeframes page
so that the task reducer is populated as well as the old task store.