Redux: only allow for a single selected tag
Until there's a use case for multiple selected tags, no reason to have
the extra logic to account for it.

This also fixes an issue where clicking on the tag didn't actually work,
so there's that... There is still an issue when going directly to a
particular tag via the url, e.g.: `/home` will not have the `home` tag
selected.