SmartTags: add isActive smart check
This creates a `rules` array where we will be able to specify a list of
rules to check against to see if a task fits a tag. The base cases are
when a tag is directly associated with a task and when a task is active,
so added those. Follow ups will be to add rules for un-tagged and
un-estimated tasks (and also fix bug where those tags don't work...).