Redux: add shim Link component
This adds a new `Link` container component to wrap the class by the same
name from React Router. I'm working to remove React Router in favor of
an approach that uses redux to store our routing information, so this
mocks up the interface I want to have for link components with that
approach.