Redux: remove react-router
This pulls the trigger. Removed React Router in favor of a redux store
based approach. I removed the Redux Router `Link` component usage as
well as a reference to `history` in `BulkTasksNew`. Added some tests to
cover a bug with the React Router approach where if you navigate
directly to a URL, it wouldn't show the appropriate tasks. Putting it in
the store instead handles this.