Introduce connectWithScratch component
This introduces a new React "Higher Order Component" that wraps a given
component with "scratch" functionality. Often when working with forms in
React, there's a question as to whether the state of the form ought to
be managed using local `setState` or going through the trouble of
putting it in Redux, which often creates some headache in terms of
managing it. If you put it in Redux, where do you put the information?
Should you clean it up after your component unmounts? How do you
initialize it?

This scratch HOC automatically handles lifecycle events, creating a spot
in the redux state for you to use as scratch to store any temporary
information for a given component. When a component is unmounted, the
scratch space is removed. It requires a unique scratch key for each
component that is mounted, and will automatically move scratch data from
one key to another if it changes. This can be useful in cases such as
when you're waiting for an ajax request to finish loading a record.

This also adds a scratch reducer and action creators.