SmartTags: select tags that have matching task
This changes the tag selectors to find tags based on whether they have a
tag that matches any of their smart rules, rather than using the static
`unfinishedTasksCount` from the server. The old count is still used on
the tag buttons and probably other places, so more bugs to fix.