Redux: move notification creation to action_creators
Rather than instantiating notifications inside the
`NotificationCheckbox` component, this moves the logic to manage that to
`notification/action_creators`. This will allow for having an interface
for managing notifications that is similar to other action creators, as
well as allowing us greater control over validation of when a
notification is added or removed.