React: fix bug with notification
The notification wasn't setting an initial value when the user had
notifications enabled. We were passing a function reference through to
the `value` prop instead of the actual return value of the function
call. Also turns out we needed to be using `checked` instead of `value`
and `onChange` instead of `onClick`.