React: introduce ReactModal for help modal
`ReactModal` seems to behave pretty well out of the box across screen
sizes. One weird behavior is that they implicitly style the modal for
you _unless_ you pass it a `className` and then you have to handle it
all yourself. So instead of handling it all myself, I used their class
names in the CSS for now. It's also possible to pass down a `style` prop
with additional custom styles, but I'm still doing things the old
fashioned way.

https://github.com/reactjs/react-modal