ChildTasks: fix bug where tasks were wrongly associated with parent
The `parentTaskId` in the task state was persisting after navigating to
a new page, so tasks added at the top level on the index page or on the
focus page ended up associated with the most recent task show page to
have been visited.

This clears out the task meta state after the component unmounts. I
tried using `connectWithScratch`, but that turned out to be prohibitive.
I'll need to convert over `NewTaskForm` to use it instead.