EditTask: change editable title to always be input
I didn't care for the interaction or usability of the way this was
working, so I changed it some. Now it's always an input, though without
a border when not focused, so it looks like just a regular heading.
Inspected Trello for some inspiration. I still might introduce an edit
icon if this turns out not to be discoverable enough. This *might* also
be better for accessibility, but that's just speculation. I was able to
re-enable the `autofocus` eslint rule, so props to me there.