ESLint: enable `react/forbid-prop-types` rule
This forbids generic prop types and requires a shape or better. Also
needed to disable the `no-typos` rule due to the linked issue. The
`shape` helper is actually not that useful, but the next minor version
of `prop-types` provides `exact` instead, which should be much better.
Will upgrade that soon.