clean up some lingering tag matchers
Missed a couple when I added it. I could see these maybe resulting in
inconsistent specs due to Capybara wait behavior, so if that happens I
will probably want to add a `have_no_tag` matcher that uses
`has_no_selector?` to properly wait for it to disappear. Let's see how
it goes.