Matchers: create HaveNoTaskWithTitle spec matcher
There is a timing issue when using `not_to have_task('blah')`.
Basically, Capybara query methods like `has_selector?` wait until a
thing is true and return as soon as that is the case, or return `false`
after some wait period. This can be a problem if the thing is in the
page and you want to wait until it disappears, say, after an ajax
request. It will return `true` right away and fail your expectation. For
this we have the negative `has_no_selector?` query method, which will
wait until that thing is *not* present and return `true`.

For the current code this isn't a problem, because it's either there or
not, but in upcoming stuff I'm working on I started to run into these
timing issues, so this adds a negated matcher that accepts a title.