EditTask: in-line editing on tasks/index page
Pretty straightforward, I just swapped in the `TaskEditTitleForm` and
`TaskEditIcon` on the task row. I'm noticing that I have to do a lot of
passing `keyPrefix` around for the scratch space, so may want to rethink
that. One option would be to assign a unique id to each component and
use that with the key.

I was able to change feature tests to use the `have_task` matcher as
they all match the same format now. I slightly changed the format,
though, to not look specifically in the `#task` form. Also, added a new
`task_row` finder for convenience and consistency. Will hopefully pull
out some more of these sorts of things soon.