ChildTasks: add leaf/root filters to list page
This adds "ROOT" and "LEAF" filters to the tasks/index page. I added a
new selector `getPartitionedTasksForRoute` and `LinkSet` component to
render a list of links with one of them being tracked as active. I
considered wrapping this in a container, but opted to leave that off for
now. It's also possible to potentially combine the 3 `tasks/` routes
into one, but opted not to do that for now. Might make sense to add as
filter params instead of as part of the url directly.