Helpers: add new `grab` helper
`grab` behaves similar to `Hash#fetch` in ruby, throwing an error if a
key is not present in the passed object. This should help shorten stack
traces so that we don't have to find our way back when something is
`undefined` somewhere unexpected. It doesn't allow passing a callback
the way the Ruby implementation does, but that may be a future
enhancement.