ESLint: disable `react/jsx-no-bind` rule
And delete `stopPropagation` helper that was only used in one place. I
think the rule is maybe a little misleading in that there are no clear
performance benefits to it by default unless used with `PureComponent`,
and even then the performance benefits may be questionable anyway. We
should probably focus on writing good looking code, and worry about
performance optimizations as a separate concern.

Thought this was a pretty decent article on the topic.
https://cdb.reacttraining.com/react-inline-functions-and-performance-bdff784f5578