EditTask: switch to textarea for task title
Using an `input` tag caused Questlog to be unusable on mobile, as the
text didn't wrap. When the title was longer than a couple of words, it
would disappear behind the other symbols on the page, so you would have
to click into the field and scroll it to see the whole thing. This was
also a problem on desktop for longer task titles.

This uses the `react-textarea-autosize` library to automatically
increase the height of the input as you type. I tried doing it myself
initially, but turns out there are a lot of complexities to auto-sizing
textareas, so the library helps out quite a bit. I needed to add an
additional `onKeyPress` callback, since `'Enter'` on `textarea` doesn't
submit the form the way it does with `input`.

Also, updated test matchers and a couple of places I still managed to
miss using `have_task` matchers. The JS tests had to change to use
`shallow` instead of `mount`, as the library throws an error with
`mount`. Probably better to just be using `shallow` anyway for
performance.

https://github.com/andreypopp/react-textarea-autosize/issues/177