Loading: generate tag priority client-side
This adds a new `getTagMetaInfo` selector to calculate the priority for
a tag client-side based on the priority of associated tasks. It will
probably be enhanced soon-ish to also take into account other tag info.
These are more steps to improve the loading experience for users.

I could see at some point adding a process to update the tag directly
instead of having meta info on the side, similar to the way `subTasks`
are set up on tasks. This just seemed like the most expedient approach
to decoupling from the server.