Loading: introduce `getTasksByTagId` selector
This is in preparation for improving both performance and usability
around server requests. Right now when there is a request to the server,
it shows a big ugly `Loading...` message. It can also be very slow once
there are a couple of hundred tasks. This is due to the fact that we
re-fetch *all* of the tasks any time one of them changes. We still have
some tight coupling to the server for things like sorting and filtering,
so we'll need to add some logic to the client side to make it possible
to make it more flexible.