Specs: clean up task editing in specs
No more need to navigate to another page, as there's an editable field
on the current page now. For some reason the blur wasn't working, as now
it ended up clicking on the Postpone button, so I specifically click on
the inert `.flashes`. Also, added the `.native.clear` bit to the
`edit_task` helper. The tests pass without it, but the title gets
jumbled still.