Specs: tidy up edit tests a tiny bit more
It turns out you can input a `"\n"` to trigger an Enter inside a
capybara input, so this does that.

Moving the tag down in `tasks_spec` is tangentially related. I'm trying
to reproduce a bug where editing a task double-submits back to the
server. It looks like I won't be able to reproduce it in feature test
form, though, as it may require a multi-threaded environment, and I
think the feature tests run on a single-threaded server.