EditTask: fix for double submit issue
`EditTaskTitle` was triggering double requests to the server on submit.
Essentially, if you hit "Enter" after editing, what would happen is that
submit triggers `blur` on the input, which again triggers a submit. This
prevents submitting while already in the process of submitting.
Unfortunately, I wasn't able to come up with a way to test this in
features, so settled for JS tests instead.

I discovered this when I added a tag to a task and two tags were
created. In most other cases the double submit would actually have no
impact. I'll also be adding a unique constraint on tags to prevent
creation of duplicates.