Loading: move task count generation to client side
This counts tasks on the client instead of relying on the count stored
on the server. I can remove the server logic soon, but want to make sure
this settles first. There may be a day in the future where it's not
sustainable to pass all of the tasks to the client and do this aggregate
logic there, but for now it makes my life easier.