Loading: don't link tasks in reducer
I ran into an issue when trying to update tasks where the top-level task
would get updated, but the one stored in `subTasks` had become stale.
The lesson seems to be that we should minimize references to an object
that are stored, otherwise we end up needing to update different
locations with the same information. We may be better served by using
selectors for data like this, maybe similar to the tag `metaInfo`, or
somehow munging tasks together in the selectors.