Selectors: add tasks directly on tag
This introduces a `processTag` function inside of our tag selectors
which adds tags directly to the tag. With some potential caveats, this
seems like it will result in the best usability inside of our
components, being able to refer to `tag.tasks` instead of needing to
import a separate selector. The code inside the selectors seems like
it'll also be a little more clear.

The caveats are that it may result in circular dependency issues if we
were to, say, want to also be able to say `task.tags`, so we'd have to
find a workaround for that. Also, this seems like it probably consumes
somewhat more memory than the previous solution, copying the tag data to
a new object, but that seems unlikely to matter here. Maybe with the
tasks a similar approach will start to matter, but I'm willing to learn
that lesson the hard way, for science!