Selectors: move sub-task logic to task selectors
Rather than handling this in the reducer, where updating an individual
task would mean extra logic in the reducer to make sure associations are
still in sync. In the selectors, it will be automatically invalidated
when tasks change and cause the data to be regenerated from scratch.
It's more work, and more data to be stored, but I don't have any
evidence to think this will be a problem (yet).