Loading: remove fetch from task update action
Re-fetching the tasks is a pretty intensive process and causes a lot of
annoying loading messages and unnecessary re-renders. Instead we can
simply take the task that is returned from the server and plug it back
into our state. Next up will be similar updates for create and destroy.

I also threw `debug` on the global scope for easier access. Now we can
use it without needing to first `import` it. It'll still be reported by
`ESLint` as a global, though, so we won't forget to take it out.