Loading: fix issue with tags not updating
After switching to not re-fetch tasks on update, a bug was introduced
with tag counters. Fetching tasks also fetches tags, but now without the
fetch the tags weren't getting updated, so that when a new tag was
created when updating a task, it wouldn't show up in the tags list.

This fixes the issue by returning a task's associated tags with the
update response and adding them to the store via an "upsert" action on
tags. This updates a tag if it's already existing in the store or adds
it if it's a new tag.

Cases where an existing tag was referenced were not affected, as the
selectors were already invalidated and counters were recalculated when
the task came back. New tags were just not added to the store, which
they are now.