Help: add some documentation on the What page
I got some feedback that it was hard to figure out what to do, so wanted
to start by at least adding some documentation. It's a poor substitute
for usability, but it's an MVP. I also copy/pasted some descriptions to
the docs folder so that I can re-use them later.

https://www.reddit.com/r/startups/comments/7nenwy/share_your_startup_january_2018/dsnjyd3/