Sidebar: introduce Sidebar component
This moves some of the top level application links to a new Sidebar
component on the page. It starts out collapsed on mobile, and expanded
on desktop and tablet, using the `matchMedia` browser API. It would
probably be worthwhile to extract this logic into a `MatchMedia`
component at some point.

**Notes**
- I added an `onNavigate` callback to the `Link` component to
  automatically close the sidebar on mobile when a link is clicked.

- I needed to add options for the chrome driver to maximize the window
  so that the sidebar wouldn't block the links on the left of the page.

https://developer.mozilla.org/en-US/docs/Web/API/Window/matchMedia