Sidebar: remove LinkSet class
Instead we can have the `Link` class figure out whether or not it's the
current route, which is basically why we created `LinkSet`. `Link` now
adds a `--active` class when it matches the current route. This is
assuming we can make all the changes we need in CSS, so we'll see if
that works out. Other alternatives I considered for a while was render
props or component injection, but both ended up feeling needlessly
complicated and roundabout.

I needed to add a `fetchRoute` in one test, which feels ugly, so
hopefully I can follow up with a change to use `shallow` and avoid
rendering `Link`s.