Specs: remove `shallowProvider` function
Turns out it wasn't necessary to make use of a `Provider` in this spec.
I just needed a single `dive()` to get what I wanted. The problem we're
trying to deal with here is the fact that `TaskListView` is wrapped in a
`dragDropContext`, so we need to dig one layer down to get at the actual
component. `dive()` does this for us.

http://airbnb.io/enzyme/docs/api/ShallowWrapper/dive.html