Sidebar: fix className duplication in Link
There are a number of places where we pass in `className` already and
wouldn't want it to be duplicated if the route happens to be active. For
example, `tag/components/button` passes in a bunch of classes, and we
don't rely on the `--active` behavior. However, as-is, it'll take
classes like:

```js
'button btn btn-default priority-1-btn'
```

and turn it into:

```js
'button btn btn-default priority-1-btn button btn btn-default priority-1-btn--active'
```

This fixes the issue by using a `baseClass` arg for the case of adding
the `--active` bit.