Errors: add `.well-known/assetlinks.json`
I've been seeing errors caused by Googlebot trying to hit the URL:

https://www.questlog.io/.well-known/assetlinks.json

Apparently this is some generic protocol to tell browsers which native
apps are associated with your site, so that, for example, a link can
open in a particular mobile app instead of in the browser. I don't have
any apps, but I want to suppress the error, so figured I'd just add an
empty file for now to suppress the routing errors. Maybe someday we'll
add links to a nice mobile app.

https://www.samclarke.com/the-well-known-directory/
https://developers.google.com/digital-asset-links/v1/getting-started