Sidebar: collapse main content with sidebar
This makes it so that the sidebar squeezes the main body when the window
shrinks to the point where they would overlap. It does not impact mobile
screens smaller than 600px width. It's a bit of a hacky solution, as the
main content div is outside of React, so we look it up and add a
`.sidebar-open` class to it, paired with a media query to set the left
margin to `250px` instead of `auto` when the screen width is < 1300 and
> 600. We should be able to clean it up some whenever we get around to
converting the rest of the app to React, where we can use global state
in redux to apply the class.