Sidebar: move `visit '/'` to spec_helper
Both a convenience and a necessity. The sidebar feature specs don't work
on CircleCI for some reason. It looks like the sidebar is overlapping
the main page, despite my attempts to resize the window. So the new
solution to this problem is to close the sidebar immediately for each
spec, and only open it as necessary to access the links. More to come.

I also found a way to squeeze the main display down when the sidebar is
open on mid-sized screens, which will be forthcoming, but for some
reason it does not work on Chrome headless.