Sidebar: toggle sidebar off at beginning of specs
This hides the sidebar immediately as specs start. I ran into some
flakiness issues with the animation, so added a script to disable them
for now. It'll speed up tests a bit, too, so bonus. If we decide at some
point we don't like having animations disabled, we can maybe integrate
`react-css-transition-group` to add and remove classes that we can
better assert against instead.

Hopefully, this should fix the build on CircleCI.