TreeView: create basic tree view for tasks
This adds a `TreeView` and `Tree` component that show tasks in a nested
list structure. It's pretty basic at this point, just displaying and not
offering any sort of options to edit or manage them. Just want to get
something live and then continue to refine it from there. I added my
handy new `AlphaModal` to make sure users are under no illusions.