SubTasks: separate route from task selectors
This moves the logic for filtering tasks based on route up to the
`list_view` container. Instead now I've exported 3 different selectors
for all, leaf, and root tasks. This should somewhat improve performance,
as now these selectors will only be cache busted when tasks change, as
opposed to having a single selector that is cache busted each time the
route changes.

**Why**

In the interest of performance, I'm trying to improve the task
selectors. Right now the `/tasks` page is super slow to load once you
have a couple of hundred tasks. The ultimate goal will be to switch from
setting `subTaskIds` on each task to instead adding a
`getTasksByParentTaskId` selector. This seemed like a pre-requisite at
the time.