Upgrades: silence help modal warning
`react-modal` introduced a new check to make sure an "app element" is
specified so that it can be hidden for accessibility when the modal is
open. I needed to add it to the default test html as several tests cause
the modal to be rendered and it complains about the missing element
otherwise.

https://github.com/reactjs/react-modal#app-element
https://github.com/reactjs/react-modal/issues/576