Modals: make more generic ModalLink component
In order to re-use this component, I generified it some. I have another
modal I'd like to add soon and found myself basically copy/pasting the
whole thing. In addition, this handles some edge cases I hadn't
accounted for. In particular, I needed to make sure only one modal could
be showing at once, so I added a `modalId` property to the `ModalLink`
component to specify exactly which modal component is open. It turns out
right now on master when the help modal is opened on the Focus page,
because there are two links to it, two modals open. It doesn't become an
issue until we have different modals, since the last one will clobber
the first.

Another issue I resolved was the fact that clicking a link on the modal
would trigger an event to propagate up to the `button` and cause it to
toggle. Moving the `Modal` component out and wrapping both in a `span`
fixed this. It still doesn't work perfect, as linking another modal from
a modal doesn't seem to work, but we can deal with that on a day where
we actually want to do that.