TreeView: add checkboxes to `TreeView`
This allows a user to mark a task complete on the tree view. Yay! A
couple of next things I'm hoping to do are to show completed tasks until
the user refreshes (and add a link to "Clear Completed"), and also allow
for marking a parent task complete with all children.