SCSS: separate SCSS out into more files
Trying to group it by component/view. It looks like there are some
classes in `common.scss` that are probably not used, so should probably
revisit. Also, more of the global styles should probably be moved from
an element to a class, such as `h1`, `h3`, and `ul` top level styles.
Aside from avoiding needless overrides, it's easier to grep for where
classes are used.