Crawlers: add `ads.txt` file
Got an error from a crawler looking for this file, so adding a blank
one. Pretty sure we'll never need this, unless we start selling ads or
something, but it's a quick way to avoid an error. Might be worth coming
up with a routing based solution at some point.

https://iabtechlab.com/ads-txt-about/

Also made `assetlinks.json` proper JSON format.