Stats: add some stat classes
The idea behind this is to be able to give the user a range of expected
productivity, rather than a static `median_productivity`. Not really
working on this right now, though, so it'll be a while until it's
integrated, but it was educational. Actually, something that I learned
from this exercise is that standard deviation can be *greater* than the
mean, which kind of blows my plans out of the water. I was hoping to
show a range to the users of:

a) what they can probably get done (mean - standard deviation)
b) what they might be able to get done (mean + standard deviation)
c) what they probably can't get done (> mean + standard deviation)

But if standard deviation is greater than mean, then what they can get
done would be negative, so that's no good. I'll probably have to do some
studying up on predictive modeling to figure out a more robust model. Or
come up with some clever fudges to make it work.

https://en.wikipedia.org/wiki/Predictive_modelling