Upgrades: fix rubocop Includes (#44)
A recent rubocop version changed the behavior of the `Includes` setting
to overwrite rather than merge options, so a bunch of files weren't
being checked for a little while. All of the files in `Includes` are in
the defaults anyway, so this should be fine to remove. Not sure why we
added the setting in the first place.

I disabled the `Rails/BulkChangeTable` setting as it makes it harder to
write reversible migrations and I've never felt any pain with migrations
being too slow.

https://github.com/rubocop-hq/rubocop/issues/6021