TypeScript: separate out tsconfig-jest.json
Ran into some complaints trying to deploy, as TypeScript was trying to
compile test files and looking for test libraries, like `enzyme`, which
were in `devDependencies`, so not installed on production. I don't love
this solution, but hopefully this will resolve the issue. Will follow up
if I hear a better approach from [this issue][1].

[1]: https://github.com/rails/webpacker/issues/1685