TypeScript: noImplicitAny in notification folder
This fixes a bug in `application.tsx` where it was passing the wrong key
to remove the notification on page unload. I made the type of
`NotificationState` have a garbage key because `{}` allows anything to
be set. Once TypeScript reports some issue with the existing structure,
I'll change it to something more appropriate.