CI: set up workflows on CircleCI
This separates out the tests into different dependent workflows on
CircleCI. If the lints don't pass, nothing else will run, and so forth,
based on the `requires` links. It's not super-optimized, as we load the
database for *every* workflow when we don't need it for a couple, and
likewise for caching and restoring dependencies. The build still doesn't
take terribly long, though, so should be fine for now.

We probably also ought to come back and set up the [`store_test_results`
piece for build insights][1]. Seems like we can use that for almost any
of the build jobs as long as they have a JUnit formatter.

[1]: https://circleci.com/docs/2.0/collect-test-data/