TypeScript: make state keys required
This updates the tests to use properly shaped state. It also adds `SET`
keys to a handful of state reducers, which I have mixed feelings about,
since it's only for the convenience of the `makeState` factory. Maybe
there's a better way to accomplish this, but this seemed like the
cleanest.