NoReact: fix task page for direct navigation
Right now the task/show page is broken when you navigate directly to it
in the URL. It depends on JavaScript setting the current task on another
page before navigation, which is kind of gross. This fills in with a
`nullTask` while the task is being loaded from the server. Not perfect,
but the goal is to replace this with a fully server rendered page at
some point, so an okay stopgap.