Specs: only freeze Date (#1766)
By default FakeTimers fakes a whole bunch of browser APIs, including
`requestAnimationFrame`, which is used heavily by Turbo. When it's
faked, it doesn't call the animation callback until you manually trigger
it via the fake clock. We only really need `Date` currently, as that's
what is used to calculate timeframes.