Deps: move fake-timers out of dev dependencies (#1769)
This is preventing deploying to production because we have a
`javascript_pack_tag` that depends on it. Kind of gross, but the quick
and dirty solution seems to be to add the dependency to the general
dependencies. Eventually, we should probably find a way to only compile
the pack tag in test mode, but not sure how to make that happen.