CircleCI: switch to Firefox for feature specs
Chrome has been super flaky, so hopefully Firefox will be better.

For some reason the Firefox driver didn't like having a heading inside
of a link, so I made it a link inside of a heading and that worked.

I also needed to change the way that I edited tasks, as it was oddly
submitting a blank task.